pyproject.toml
[project]
name = "plantpal"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "matplotlib>=3.10.8",
    "panda>=0.3.1",
    "pandas>=2.3.3",
    "path>=17.1.1",
    "pillow>=12.1.0",
    "scikit-learn>=1.8.0",
    "seaborn>=0.13.2",
    "streamlit>=1.54.0",
    "tensorflow>=2.20.0",
]
